Instagram Bot to Follow People
Is your feature request related to a problem? Please describe.
Every small business is now on instagram and one of most frustrating thing to market your product is to follow people so that they discover your page.

Describe the solution you'd like
This instagram bot uses Selenium python package to control the browse and automate the task. Version 2.0 of the bot uses InstaPy module to solve the issue, without getting your account limited.